Apply sugestions from review
